// Desafio-Frontend/src/types/types.ts
export interface VehicleInterface {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  consumables: string;
  vehicle_class: string;
}

export interface FormTypes {
  nome: string;
  email: string;
  telefone: string;
  cpfCnpj: string;
  cep: string;
  endereco: string;
  complemento: string;
  cidade: string;
  bairro: string;
  uf: string;
  pagamento: string;
  numeroCartao: string;
  validadeCartao: string;
  nomeImpresso: string;
  cvv: string;
}

export interface FormData {
  nome: string;
  email: string;
  telefone: string;
  cpfCnpj: string;
  cep: string;
  endereco: string;
  complemento: string;
  cidade: string;
  bairro: string;
  uf: string;
  pagamento: string;
  numeroCartao: string;
  validadeCartao: string;
  nomeImpresso: string;
  cvv: string;
}